// src/app/core/services/api/legacy-managent-api-client/model/action.ts
// @ts-nocheck

/**
 * EDC REST API
 * All files merged by open api merger
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Constraint} from './constraint';

export interface Action {
  type?: string;
  includedIn?: string;
  constraint?: Constraint;
}