// src/app/core/services/api/legacy-managent-api-client/model/permission.ts
// @ts-nocheck

/**
 * EDC REST API
 * All files merged by open api merger
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Action} from './action';
import {Constraint} from './constraint';
import {Duty} from './duty';

export interface Permission {
  action?: Action;
  assignee?: string;
  assigner?: string;
  constraints?: Array<Constraint>;
  duties?: Array<Duty>;
  target?: string;
  uid?: string;
}