// src/app/core/services/api/legacy-managent-api-client/model/policy.ts
// @ts-nocheck

/**
 * EDC REST API
 * All files merged by open api merger
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Duty} from './duty';
import {Permission} from './permission';
import {Prohibition} from './prohibition';

export interface Policy {
  type?: Policy.TypeEnum;
  assignee?: string;
  assigner?: string;
  extensibleProperties?: {[key: string]: object};
  inheritsFrom?: string;
  obligations?: Array<Duty>;
  permissions?: Array<Permission>;
  prohibitions?: Array<Prohibition>;
  target?: string;
}
export namespace Policy {
  export type TypeEnum = 'SET' | 'OFFER' | 'CONTRACT';
  export const TypeEnum = {
    Set: 'SET' as TypeEnum,
    Offer: 'OFFER' as TypeEnum,
    Contract: 'CONTRACT' as TypeEnum,
  };
}